Issue 1:
Fixed the pending issue of stretch view not stretching properly after rotation.

Fix solution:  The existing code of stretch simply re-used the shortcut call of resizing the right horizontal shape object and bottom vertical shape object. This was fine however the resize was made in relative to the bottom vertical shape object and not with "stretch view" at the bottom right corner. This was fixed with relative size being calculated with the stretch view.

Issue 2:

With the previous fix, there were some improper resize issues with the rotation angle between  45-90, 135-180, 225-270, 315-360. This issue was the reason as the resize reference was taken as 90,180,270 and 360.

Fix solution :
The reference angle was changed to 45, 135, 225 and 315 thereby fixing the proper working of resize at all angles.